Disable Firefox, don't restart explorer unnecessarily

+++ upgrade.ps1
@@ -1,13 +1,28 @@
 $key = 'HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced'
-Set-ItemProperty $key Hidden 1
-Set-ItemProperty $key HideFileExt 0
-Set-ItemProperty $key ShowSuperHidden 1
-Stop-Process -processname explorer
+$advanced = Get-ItemProperty $key
+
+$restart_explorer = 0
+if (-not $advanced.Hidden) {
+    Set-ItemProperty $key Hidden 1
+    $restart_explorer = 1
+}
+if ($advanced.HideFileExt) {
+    Set-ItemProperty $key HideFileExt 0
+    $restart_explorer = 1
+}
+if (-not $advanced.ShowSuperHidden) {
+    Set-ItemProperty $key ShowSuperHidden 1
+    $restart_explorer = 1
+}
+
+if ($restart_explorer) {
+    Stop-Process -processname explorer
+}
 
 choco upgrade -y cmake
 choco upgrade -y conan
 choco upgrade -y cppcheck
-choco upgrade -y firefox
+#choco upgrade -y firefox
 choco upgrade -y git
 choco upgrade -y llvm
 choco upgrade -y mingw
